+++ Sheepskin/Library/Data/Instrument/CryptoCurrency/CryptoCurrency.h
@@ -11,7 +11,7 @@
 class CryptoCurrency: public Instrument {
 public:
     explicit CryptoCurrency(std::string id);
-    long double getVolatility(const Record &record) override;
+    double getVolatility(const Record &record) override;
     double getPrice(const Record &record) override;
 };
