+++ Sheepskin/Library/Strategies/FirstStrategy/FirstStrategy.h
@@ -24,16 +24,16 @@ class FirstStrategy: public Strategy {
 public:
     explicit FirstStrategy(Instrument* instrument);
     StrategyResult eval() override;             // zwracać ma wynik w postaci rise_chance, fall_chance, maintance_chance
-    std::vector<long double> listOfTangens();
-    long double calculateTangens(const Record& record1, const Record& record2);
-    int lookForChange(std::vector<long double> values, int index);     // zwraca pozycję w wektorze gdzie nastąpiła zmiana
-    long double calculateDifference(int down, int up);                // zwraca różnicę w cenach między ostatnimi dwoma zmianami
-    long double sumOfDifference(std::vector<long double> tangens);  // zwraca sumę wszystkich zmian
+    std::vector<double> listOfTangens();
+    double calculateTangens(const Record& record1, const Record& record2);
+    int lookForChange(std::vector<double> values, int index);     // zwraca pozycję w wektorze gdzie nastąpiła zmiana
+    double calculateDifference(int down, int up);                // zwraca różnicę w cenach między ostatnimi dwoma zmianami
+    double sumOfDifference(std::vector<double> tangens);  // zwraca sumę wszystkich zmian
     double setPrecision(double value, int precison);
     double calculateMovingAverage();
-    std::tuple<int, int, int>getNumberOfStatus(std::vector<long double> tangens);   // zwraca liczbę ile razy się cena podnosiła, ile malała i ile łacznie się zmieniałą
+    std::tuple<int, int, int>getNumberOfStatus(std::vector<double> tangens);   // zwraca liczbę ile razy się cena podnosiła, ile malała i ile łacznie się zmieniałą
     std::tuple<double, double, double> calculateChances();                          // oblicza % na podstawie ilości zmian oraz ilości różnicy cen oraz obecnym trendzie
-    Price status(long double value);                                    // na podstawie wartości tangensa określa czy ceny maleją czy rosną
+    Price status(double value);                                    // na podstawie wartości tangensa określa czy ceny maleją czy rosną
     Price lastStatus();                                                 // zwraca ostatni stan > czy jest malejący, stały czy rosnący
 };
